Change names of config files, use java config instead of annotations
