--- src/main/java/com/springframework/config/GreetingServiceConfigurationOtherPackage.java
@@ -1,13 +1,0 @@
-package com.springframework.config;
-
-import guru.springframework.sfgdi.services.ConstructorGreetingService;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class GreetingServiceConfigurationOtherPackage {
-    @Bean
-    ConstructorGreetingService constructorGreetingService(){
-        return new ConstructorGreetingService();
-    }
-}
