+++ src/main/java/guru/springframework/sfgdi/config/GreetingServiceI18nServiceConfiguration.java
@@ -1,0 +1,28 @@
+package guru.springframework.sfgdi.config;
+
+import guru.springframework.sfgdi.services.I18nEnglishGreetingService;
+import guru.springframework.sfgdi.services.I18nSpanishGreetingService;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.stereotype.Service;
+
+@Configuration
+public class GreetingServiceI18nServiceConfiguration {
+
+    @Profile("EN")
+    @Bean("i18nService")
+    I18nEnglishGreetingService i18nEnglishGreetingService(){
+        return new I18nEnglishGreetingService();
+    }
+
+    @Profile({"ES", "default"})
+    @Bean("i18nService")
+    I18nSpanishGreetingService i18nSpanishGreetingService(){
+        return new I18nSpanishGreetingService();
+    }
+
+
+
+
+}
