--- src/main/java/guru/springframework/sfgdi/services/I18NSpanishService.java
@@ -1,16 +1,0 @@
-package guru.springframework.sfgdi.services;
-
-import org.springframework.context.annotation.Profile;
-import org.springframework.stereotype.Service;
-
-/**
- * Created by jt on 12/27/19.
- */
-@Profile({"ES", "default"})
-@Service("i18nService")
-public class I18NSpanishService implements GreetingService {
-    @Override
-    public String sayGreeting() {
-        return "Hola Mundo - ES";
-    }
-}
