+++ src/main/java/guru/springframework/sfgdi/services/I18nEnglishGreetingService.java
@@ -1,13 +1,5 @@
 package guru.springframework.sfgdi.services;
 
-import org.springframework.context.annotation.Profile;
-import org.springframework.stereotype.Service;
-
-/**
- * Created by jt on 12/27/19.
- */
-@Profile("EN")
-@Service("i18nService")
 public class I18nEnglishGreetingService implements GreetingService {
     @Override
     public String sayGreeting() {
