+++ src/main/java/guru/springframework/sfgdi/services/I18nSpanishGreetingService.java
@@ -1,0 +1,8 @@
+package guru.springframework.sfgdi.services;
+
+public class I18nSpanishGreetingService implements GreetingService {
+    @Override
+    public String sayGreeting() {
+        return "Hola Mundo - ES";
+    }
+}
