+++ src/main/java/guru/springframework/sfgdi/services/PrimaryGreetingService.java
@@ -1,13 +1,5 @@
 package guru.springframework.sfgdi.services;
 
-import org.springframework.context.annotation.Primary;
-import org.springframework.stereotype.Service;
-
-/**
- * Created by jt on 12/27/19.
- */
-@Primary
-@Service
 public class PrimaryGreetingService implements GreetingService {
 
     @Override
