+++ src/main/resources/application.properties
@@ -1,3 +1,3 @@
-spring.profiles.active=EN,dog
+spring.profiles.active=ES,cat
 spring.profiles.default=default
